+++ source/extensions/tracers/opentelemetry/samplers/always_on/always_on_sampler.h
@@ -26,7 +26,7 @@ public:
                               const std::set<SpanContext> links) override;
   std::string getDescription() const override;
 
-  std::string modifyTraceState(const std::string &/*span_id*/, const std::string &/*current_trace_state*/) const  override { return {}; };
+  std::string modifyTracestate(const std::string &/*span_id*/, const std::string &/*current_tracestate*/) const  override { return {}; };
 
 private:
 };
