+++ source/extensions/tracers/opentelemetry/samplers/dynatrace/BUILD
@@ -24,7 +24,7 @@ envoy_cc_extension(
 envoy_cc_library(
     name = "dynatrace_sampler_lib",
     srcs = ["dynatrace_sampler.cc"],
-    hdrs = ["dynatrace_sampler.h", "trace_state.h",],
+    hdrs = ["dynatrace_sampler.h", "tracestate.h",],
     deps = [
         "//source/common/config:datasource_lib",
         "//source/extensions/tracers/opentelemetry/samplers:sampler_lib",
