+++ source/extensions/tracers/opentelemetry/samplers/dynatrace/dynatrace_sampler.h
@@ -28,7 +28,7 @@ public:
                             
   std::string getDescription() const override;
 
-  std::string modifyTraceState(const std::string &span_id, const std::string &current_trace_state) const  override;
+  std::string modifyTracestate(const std::string &span_id, const std::string &current_tracestate) const  override;
 
 private:
   std::atomic<uint32_t> counter_;
