+++ source/extensions/tracers/opentelemetry/tracer.h
@@ -133,7 +133,7 @@ public:
 
   ::opentelemetry::proto::trace::v1::Span::SpanKind spankind() const { return span_.kind(); }
 
-  std::string tracestate() { return span_.trace_state(); }
+  std::string tracestate() const { return span_.trace_state(); }
 
   /**
    * Sets the span's tracestate.
